FRONTEND-28: move shared row styling to /styles/row.module.css
