--- src/app/apps/[slug]/edit/edit.module.css
@@ -1,19 +1,0 @@
-.row {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    margin-bottom: 1rem;
-    border: 1px solid #ccc;
-
-}
-.row > a {
-    text-decoration: underline;
-    margin-right: 1rem;
-}
-.row > p {
-    margin-right: 1rem;
-}
-
-.row > button {
-    margin-right: 1rem;
-}
